+++ src/components/button-load.js
@@ -1,0 +1,4 @@
+export default () => {
+  return `
+    <button class="load-more" type="button">load more</button>`;
+};
